move model file to a dedicated folder
