+++ internal/api/front/perses/src/app/project/dashboard/dashboard-details/dashboard-details.component.ts
@@ -16,7 +16,7 @@ import { DashboardService } from '../service/dashboard.service';
 import { ToastService } from '../../../shared/service/toast.service';
 import { ProjectService } from '../../project.service';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { DashboardModel } from '../dashboard.model';
+import { DashboardModel } from '../model/dashboard.model';
 
 @UntilDestroy()
 @Component({
