+++ internal/api/front/perses/src/app/project/dashboard/dashboard-details/dashboard-variables/dashboard-variables.component.ts
@@ -12,9 +12,9 @@
 // limitations under the License.
 
 import { Component, Input, OnInit } from '@angular/core';
-import { DashboardVariable } from '../../dashboard.model';
+import { DashboardVariable } from '../../model/dashboard.model';
 import { DashboardFeedService } from '../../service/dashboard-feed.service';
-import { VariableFeedRequest } from '../../dashboard-feed.model';
+import { VariableFeedRequest } from '../../model/dashboard-feed.model';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { ToastService } from '../../../../shared/service/toast.service';
 import { EventFeedService } from '../../service/event-feed.service';
