+++ internal/api/front/perses/src/app/project/dashboard/model/dashboard-feed.model.ts
