+++ internal/api/front/perses/src/app/project/dashboard/service/dashboard-feed.service.ts
@@ -14,7 +14,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { ErrorHandlingService } from '../../../shared/service/error-handling.service';
-import { VariableFeedRequest, VariableFeedResponse } from '../dashboard-feed.model';
+import { VariableFeedRequest, VariableFeedResponse } from '../model/dashboard-feed.model';
 import { Observable } from 'rxjs';
 import { UrlBuilderUtil } from '../../../shared/utils/url-builder.util';
 import { catchError } from 'rxjs/operators';
